Ok, that wasn't the problem. Submodules was.
Submoduling apparently kind of happens whenever you clone a repo into a
repo. It doesn't seem to get set up correctly, though, so it just breaks
things. For now, I'm going to have to just install those modules as just
files, but I think in the future the subtree stuff might be worth
investigating.

I've also fixed the permissions on things. I kinda made a mess of that.